+++ sdk/formrecognizer/ai-form-recognizer/samples/javascript/recognizeReceiptFromUrl.js
@@ -20,7 +20,7 @@ async function main() {
     "https://raw.githubusercontent.com/Azure-Samples/cognitive-services-REST-api-samples/master/curl/form-recognizer/contoso-allinone.jpg";
 
   const poller = await client.beginRecognizeReceiptsFromUrl(url, {
-    includeTextDetails: true,
+    includeTextContent: true,
     onProgress: (state) => {
       console.log(`analyzing status: ${state.status}`);
     }