+++ sdk/formrecognizer/ai-form-recognizer/samples/typescript/src/recognizeReceiptFromUrl.ts
@@ -20,8 +20,8 @@ export async function main() {
     "https://raw.githubusercontent.com/Azure-Samples/cognitive-services-REST-api-samples/master/curl/form-recognizer/contoso-allinone.jpg";
 
   const poller = await client.beginRecognizeReceiptsFromUrl(url, {
-    includeTextDetails: true,
-    onProgress: (state: BeginRecognizeReceiptPollState) => {
+    includeTextContent: true,
+    onProgress: (state) => {
       console.log(`analyzing status: ${state.status}`);
     }
   });